vestir-prototype/embedding-sidecar/requirements.txt
# PyTorch: install per platform from https://pytorch.org if needed, or:
torch>=2.1.0
transformers>=4.44.0
accelerate>=0.26.0
safetensors>=0.4.0
pillow>=10.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
# Marqo FashionSigLIP (open-clip backend)
open-clip-torch>=2.24.0
ftfy>=6.1.0
sentencepiece>=0.2.0
timm>=1.0.0
